cmd/snap-confine: remove stale mount profile along stale namespace
When snap-confine attempts to start up and notices that the mount
namespace is stale and can be discarded it does so correctly but it
doesn't remove the .fstab file that has become stale because the mount
namespace was discarded. When snap-update-ns is subsequently used it
thinks that the mount namespace has all the changes applied to it and
then does nothing.

Thanks to Alan Pope for reporting the issue and providing helpful data.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>