snap: add ReadInfoExceptSize
The snap-{run,confine,exec} pipeline culminates with snap-exec reading
the info of the snap that needs to be executed and running the appropriate
command. If the snap in question is a try mode snap in /tmp or other path
that is not accessible from the execution environment the code in ReadInfo
falls over inability to compute the utterly useless Size field.

Perhaps the code in snap-exec should be more optimized to be as minimal
as needed but for now let's introduce a new ReadInfo variant that skips
the size computation.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>