tests: adjust regression test for broken snaps
When we "snap try" some directory and then move it around we are not
really breaking the snap. The bind mount keeps the contents visible
and operational. To really break a snap we must modify the structure as
expected by snapd, for example the snap.yaml file. This patch changes
the regression test to move the snap.yaml file away to see that the
broken snap detector works as it should.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>