packaging/opensuse: swap priority between GOPATH and source package
Normally there is only one copy of the code, then the order of GOPATH
and where we expect the code to be doesn't matter. In spread we build
the code that has been copied to %{_topdir} (aka /usr/src/packages/)
but we also have a copy that was ingested by spread into /home/gopath/
and set the GOPATH variable to point there. This arrangement makes "go
build" build the vanilla unpatched (mainly wrt version since there are
no patches in spread) copy.

This patch fixes it by swapping the order of the two entries in GOPATH.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>