snapctl: add "services"
This change adds the `services` verb to `snapctl`.

It includes a refactor to move `notesForSvc` from `cnd/snap` to a
method of `client.AppInfo`, and moves `clientAppInfosFromSnapAppInfos`
from `daemon` to `client.AppInfosFromSnapAppInfos`; this way we avoid
duplicating this bit of the logic.

I've discussed this with Gustavo, and there'll be a followup that
makes `services` more useful for scripts, but it's separate work (and
affects `snap services` as well).