packaging/opensuse: move most logic to snapd.mk
Packaging in spec files, debian rules, arch and all kinds of other
formats is nice when you can essentially push the bulk of the work to
"make install" or similar. In our project we don't yet have that luxury
because we sport two separate makefile systems (one autotools, one
handmade, using different libexecdir in each time for extra
eyebrow-raising) some go executables that are built with non-default
options, a small set of files that are copied manually around and two
major build modes (core vs classic-except-ubuntu) that differ in the set
of files we need to ship.

To start shaving they Yak I propose that we move most of the packaging
muscle in openSUSE's snapd.spec to snapd.mk.

Where the spec file is ugly is around conditional handling. I was unable
to find a way to have RPM handle bot the %if expr conditional as well as
the %{?foo} and %{!?foo} conditionals sanely. Fortunately the makefile
world is much saner and this is where most of the logic now lives.

The spec file builds correctly on openSUSE Leap 42.3, 15.0 and
Tumbleweed. Spread testing against Leap 42.3 suggests that I didn't
break anything either.

The spec file is much more precise now, claiming ownership of many more
files and directories belonging to snapd that are created dynamically.

I tried to document the makefile and have belts-and-suspenders approach
so that nothing (or as little as I could) would fail silently. I haven't
yet tried using it on Debian. I would attempt to use it for Arch next,
where it should be easiest, and then on Fedora and Amazon Linux that
share packaging.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>