daemon: spool sideloaded snap into blob dir (#6093)
* daemon: spool sideloaded snap into blob dir

Before this change a sideloaded snap would get written out to /tmp
before being copied into /var/lib/snapd/snaps. As these are often
separate devices this means sideloading a snap would be unnecessarily
slow, and use up a potentially large amount of /tmp (which might be
small and in-memory, on core).

This change makes it so that the snap is instead written to a
temporary (and hidden) file in /var/lib/snapd/snaps directly.

* osutil, overlord/snapstate: remove old sideload files from snapstate.Ensure

* overlord/snapstate: make the test a little quicker

* dirs, daemon, overlord/snapstate: move the sideload blob temp prefix def to dirs

* daemon, dirs, osutil, overlord/snapstate: address review feedback

This is mosty a rename and a small change to only run the cleanup code
once a day.

Thanks @pedronis for the excellent feedback.