cmd/snap-discard-ns: add support for --from-snap-confine
The snap-confine program delegates some work to external programs such
as snap-update-ns. When doing that it passes an extra option that indicates
the program is invoked by snap-confine and should not perform its own
locking.

We are about to start calling snap-discard-ns this way so it needs to
grow a command line option to express this.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>