wrappers: remove all desktop files from a snap on removal
When we added support for parallel installs we broke removal of
some desktop files. This was reported in LP: 1791182. The problematic
commit is 84008e.

This PR changes the way we write and remove desktop files. One
problem here is that desktop files of a foo_instance will get
removed when "snap remove foo" happens. Given that parallel installs
are experimental this might be acceptable (or not).