features: add module for querying enabled snapd features
This builds upon the ability to export enabled experimental features to
userspace. We have the Go side for controlling features via "snap set
core experimental.foo 1" and the C side for querying them but we still
need the Go counterpart. This patch adds one.

The API is extremely simple, consisting of a new type, SnapdFeature,
along with a IsEnabled method that performs a stat call.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>