daemon: update activation handling to support API changes in github.com/coreos/go-systemd/activation
The github.com/coreos/go-systemd/activation package was bumped to v17 which
introduced an API change, activation.Listeners(unsetEnv bool) was changed to
activation.Listeners(), where it defaults to unsetEnv = true.

Refactor the activation to use activation.Files() which has not changed across
package versions and allows to preserve the original intention of the code in
daemon.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>