client, cmd, etc: move Notes and AppInfo×AppInfo into cmd
Wrt having Notes in client, @pedronis pointed out that

> Notes as a concept belong to the CLI UX, from that POV even
> clientutil seems the wrong place. It's not clear that clients (vs
> CLI UX) need notes or would agree on what they should be. Basically
> if client is about talking the snapd API from a client-side POV and
> considering we had requests to evolve client as a library vs just a
> piece of our clients/the CLI experience, it seems over-deterministic
> to put Notes under client/.

(and similar problems with `client.AppInfosFromSnapAppInfos`).

so we discussed where they would fit. `client/clientutil` was
discarded because, again, not really `client`-ish. `cmd/cmdutil`
looked reasonable, but on the other hand having a package for just two
functions was weird. So for now it's `cmd.ClientAppInfoNotes(...)` and
`cmd.ClientAppInfosFromSnapAppInfos(...)`.

HTH, HAND.