tests: reduce verbosity around package installation
Package installation is particularly verbose, not because of actual
invocation of the system package manager, but because there's a pack of
indirection layers to simplify cross distribution portability.

Recently we ran into an issue where a Travis job is killed after
exceeding log size. As the package trace logs are rarely a problem I've
made as small patch that silences shell trace output around that logic.

To maintain minimal trace we now echo the set of packages to install
along with the name of the function that was invoked.

Thanks to Maciej Borzecki for the trick with "set -o"

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>