packaging: generate snapd.defines.mk from snapd.spec
Instead of having to teach snapd.mk to query the build system for
definitions and pass multiple with_ variables that define the behavior
of the build system we can delegate the responsibility of creating a
file with all the definitions to the packaging world and simply have the
build system include that.

This is what snapd.defines.mk is for. The snapd.spec file creates it and
snapd.mk includes it. There's some nice reduction in boilerplate and the
logic for calling snapd.mk is now much more simple.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>