tests/lib/pkgdb: disable weak deps on Fedora
Fedora images we use come with a number of preinstalled package, among
them gtk3. Those packages are needed to run the tests. The
xdg-desktop-portal-gtk package uses this in the spec:

  Supplements:    (gtk3 and (flatpak or snapd))

As a result, when snapd is installed, we will unintentionally pull in
xdg-desktop-portal-gtk and its dependencies breaking tests. For this
reason, disable weak deps altogether.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>