selinux: package to query SELinux status and verify/restore file contexts
Adding a new package for working with SELinux.

There is basic functionality to query SELinux state, enabled and
permissive/enforcing. The code founds out where the selinux filesystem is
mounted and uses that path to query the status, rather than hardcoding
/sys/fs/selinux.

The package has additional helpers to verify whether a path is labeled accroding
it its default context and possibly restore the context. Trying to avoid calling
to libselinux directly, the command line tools `matchpathcon` and `restorecon`
are used.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>