httputil: retry on temporary net errors (#6287)
We currently retry network operations on various high level errors
like 500. However we do not retry on network errors like DNS
connection refused. To fix that this PR will also retry when
we hit net.Error conditions that are marked as Temporary() (like
dns connection refused errors when the DNS server is not quite
ready yet).

This has also a couple of hacks detecting error messages
directly because of go1.6 behavior and Arch ATM using the cgo
resolver. We should try to work to not need them later.

This should fix LP: #1807770