interfaces/backends: make use of UnusableAppArmor
The 2.36.x release branch was using PartialAppArmor coupled with
additional check for lack of parser features to know that apparmor ought
to be disabled inside snapd. Meanwhile in master a new level,
UnusableAppArmor, was introduced to express that. This patch reconciles
the code to be in sync with master.

While I was at it, I added a test for UnusableAppArmor level.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>