overlord/snapstate: during refresh, re-refresh on epoch bump
Without this change an installed snap that went through an epoch
change with an old gatekeeper revision would not see the newer
revision until the next auto-refresh.

In other words, if a snap had epoch 0 and the refresh found the tip of
the channel to be on epoch 1, with epoch 1* available, then without
this change the snap would only refresh to epoch 1*, and would reach
epoch 1 on the next auto-refresh.

This change makes snapd notice the epoch bump, and re-refresh the
snaps that went through it.