many: allow core as a fallback for core16
This PR allows to use the "core" snap when the "core16" snap
is required. This allows us to gently transition people to
use the core16 base without having a massive amount of extra
downloads and extra snaps on the machines.

Eventually we replace core with core16+snapd but that will
happen further down the line.