tests: deny ioctl - TIOCSTI with garbage in high bits
We were alerted to the fact that seccomp filter matching is misbehaving
when the kernel uses 32bit value for an argument that is 64bit. In other
words where the high bits are simply ignored.

The seccomp profile generated by snapd for regular seccomp profiles for
ioctl has code similar to this one:

    # filter for arch x86_64 (3221225534)
    if ($arch == 3221225534)
      # filter for syscall "ioctl" (16) [priority: 65531]
      if ($syscall == 16)
	if ($a1.hi32 == 0)
	  if ($a1.lo32 == 21522)
	  else
	    action ALLOW;
	else
	  action ALLOW;
      # default action
      action ERRNO(1);

Note that the comparison is such so that if a1.hi32 is not zero, the
action is ALLOW.

A simple solution would be to perform arithmetics fix the generated
bytecode. Alas our libseccomp and golang-seccomp bindings cannot do that
at this time.

To resolve the problem we can take advantage of the recently introduced
global profile. A global profile that rejects any attempts to use ioctl
TIOCSTI, without regard for the contents of the high word, is rejected.
The global profile contains the equivalent of the following profile, for
all architectures:

    if ($arch == 3221225534)
      # filter for syscall "ioctl" (16) [priority: 65533]
      if ($syscall == 16)
	if ($a1.hi32 & 0x00000000 == 0)
	  if ($a1.lo32 & 0xffffffff == 21522)
	    # TIOCSTI is 21522
	    action ERRNO(1);
      # default action
      action ALLOW;

This patch contains a C program that generates the blacklist as well as
a regression test for the fix that was silently landed earlier.

Architectures of a given word order are handled with one BPF program.
Specifically for little endian that is x86_64, aarch64 and ppc64le and
for big-endian only s390x. The issue does not affect any 32bit
architecture.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>