snap-confine: provide proper error message on sc_sanity_timeout
The code that communicates with the helper process does setup
a "sanity" timeout currently. If that timeout is reached then
snap-confine dies. Unfortunately the error message is misleading
and says "interrupted system call" which does not help finding
the underlying issue.

This PR changes the code so that the timeout is checked before
the read error is printed. This will lead to the slightly better:
```
error: sanity timeout expired
```
which is still not great but it helps to see that this is not
an error in the code but something with the interaction of
snap-confine and its helpers.