image,cmd/snap,tests:  introduce prepare-image --classic (#6467)
This adds support for a classic mode to prepare-image (requires --classic to be used) which is mostly a subset of the full operations needed for the core case. Given that for classic a gadget is optional for now there is no support or use for gadget unpacking. Other bits are made optional.

It also adds the capability in classic mode to seed classic snaps, while giving an error for them if that is attempted for a core image.

Gadget unpacking for this mode could be added back later via a dedicated option.