packaging: disable systemd environment generator on 18.04 (#6470)
The PATH of systemd services does not contain "/snap/bin" because services
do not read /etc/environment (that's PAM) or login.defs (getty). To fix that
we added a systemd environment generator that would modify PATH
in snapd 2.35.1 and this works successfully in 18.10+ now. Unfortunately
the systemd version in 18.04 has a bug with the handling of the PATH
with the generators which leads to an undesired side effect when the
generator is used: we get a PATH that does no longer
include "/usr/local/{,s}bin which will break existing systems.

Hence the systemd environment generator we ship must be disabled
on 18.04 because of the systemd bug #1771858. This was reported for
us as LP: 1814355.

This PR disables the generator on 18.04 (which is most unfortunate)
until we can fix systemd and adds a spread test.

A slightly simpler version of this fix was uploaded to bionic as
2.37.1.1.