cmd/snap: tweak man output to have no doubled up .TP lines
As reported in [lp:1814767], snap's manpage contains doubled up .TP
lines, which for groffs older than ~16.04 (in particular, groffs
missing [the commit] that fixes the issue) means the output is less
than ideal.

Fixing this in goflags is probably going to be tricky. I'll be trying
to do that at some point, but meanwhile this fixes the issue by
post-processing the generated manpage with a heckin' regexp.

[lp:1814767]: https://bugs.launchpad.net/snapd/+bug/1814767
[the commit]: https://git.savannah.gnu.org/cgit/groff.git/commit/?id=49abd7a9093ae903bb0032eb74d647fb481d2acb