interfaces/seccomp: regenerate changed profiles only
When setting up a seccomp backend for a snap, we unconditionally regenerate all
profiles, not just the ones that were changed. This comes with a significant
performance penalty, if a snap has many apps and hooks. Try to be smarter and
regenerate only the profiles that were changed.

Some hard numbers collected when installing test-snapd-policy-app-consumer:

|      | Before | After |
|------+--------+-------|
| RPI3 | 2m36s  | 1m24s |
| x220 | 3m24s  | 1m48  |
| PC   | 1m41s  | 39s   |

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>