debian: ensure leftover usr.lib.snapd.snap-confine is gone
In commit e16659e201d (2017-03-28, snapd v2.23.6) we renamed
/etc/apparmor.d/usr.lib.snap-confine to usr.lib.snap-confine.real
to fix LP: #1673247 - however some people (upgrades?) still have
the old usr.lib.snap-confine file. This seems to be loaded instead
of the correct usr.lib.snap-confine.real profile. To fix this we
use the rather blunt approach to remove the file forcefully if
it is unchanged.