cmd/snap-confine: track per-app and per-hook processes
When the app-refresh-awareness feature is enabled, snap-confine will
track processes in a v1 cgroup hierarchy for the pids controller. This
allows for per-app and per-hook process enumeration and tracking.

Currently nothing is using this feature, in will soon be used by snapd's
process tracking for soft and hard refresh checks, though.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>