tests: improve snaps-system-env test
In 18.04 we had a regression when we installed the snapd systemd
environment generator. Installing it caused LP: #1814355 and
the PATH for systemd services looked suddently like:
```
   /sbin:/usr/sbin:/bin:/usr/bin:/snap/bin
```
instead of the expected:
```
   /usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/snap/bin
```

A better test that checked the *full* expected PATH instead of just
the addition of :/snap/bin would have caught this regression.

This PR adds the test for the full PATH now.