daemon: allow downloading snaps blobs via .../blobs
This change exposes a new endpoint, /v2/snaps/<name>/blob that, if the
snap is installed, active, and not in try mode, allows downloading the
snap blob. As the data is the same as what could be re-created by
running mksquashfs on /snap/<name>/current/, there is no info leak.