daemon: refactor debug a little bit
This refactors debugCmd into its own file, and modifies getDebug and
postDebug to call getBaseDeclaration, instead of having getDebug call
postDebug which was just asking for trouble down the line.

Also it splits the tests into its own file, but still in the daemon
package; refactoring that will take a bit more work.