interfaces/udev/spec.go: add spec for interface to disable udev tagging
Disabling udev tagging is needed for the greengrass-support interface, as greengrass the snap needs to manage it's own device cgroups for the containers it manages, and as such, it can't currently operate in the device cgroup that would be enforced by snapd if the greengrass snap plugs an interface such as gpio, etc. which triggers udev tagging, which then triggers the device cgroup to be enforced.
Also refactoring the greengrass-support tests to test this feature, and adding the spec to greengrass-support using a helper property in the commonInterface.