apparmor: remove code that pretends to unload apparmor profiles
unloadProfiles() is currently called with a list of profile names and
it invokes apparmor_parser --remove on those names. Unfortunately,
apparmor_parser --remove requires full paths to the profiles, not just a
profile name, so the profiles were not removed. While we could prepend
dirs.SnapAppArmorDir to the profile names, by the time unloadProfiles()
is called it is too late because the profiles on disk have already been
removed. This was never caught before because apparmor_parser less than
2.13 would return 0 when given profiles that didn't exist and so the
error check would never fire.

The code removal preserves the actual current behavior for all versions
of apparmor. The previous code was somewhat misguided for the current
implementation of snapd since 'snap remove' does not guarantee there are
no processes running for the removed snap. While unloadProfiles() could
be fixed to unload the profiles, this could result in a snap process
being moved to unconfined upon 'snap remove'. Comments are added to show
how to remove a profile using sysfs but with a warning that we should
not do this until all processes are guaranteed to be killed as part of
'snap remove'.