many: remove unused code (mostly)
The mostly bit is because I couldn't resist a teeny tiny refactoring
along the way.

Anyhow, this is the result of running the `unused` static analyser on
the code. There's a bunch more, but unused doesn't know about gocheck,
so it gets confused about some things, and separating the wheat from
the chaff is tedious.