snap: remove obsolete license-* fields in the yaml
We do not use these license-{agreement,versin}. The original purpose
was to eventually show a license agreement to the user before the
snap can be installed. However we never implemented this and it seems
nowdays we would do that differently (with entitlements) anyway.

So droping these fields from the code seems to be prudent.