overlord/snapstate/backend: make LinkSnap clean up more
Without this, if an error occurred during `updateCurrentSymlinks` (or
later, but right now there's nothing 'later'), the wrappers, the data
directory for the revision, and the current symlinks, could all be
left behind. The former could leave spurious wrappers, while the
latter could confuse remove when it found an unexpectedly non-empty
directory.

This change prevents this from happening. It does not, however, clean
up situations where this has happened in the past. We could write a
state/patch for it, but deleting data because we think it's there by
accident sounds wrong to me.

There's a nastier clean-up that still needs doing:
`boot.SetNextBoot(info)` needs to get undone.