snapstate: only keep 2 snaps on classic
We had a recent discussion with the foundations team about the number
of snaps that should be kept on the system for rollback. Because
classic does not have a kernel snap the number of 2 snaps (one active
and one as a rollback target) is sufficient there. This PR changes
the default on classic from 3 to 2 and updates the tests.