cmd/snap-confine: call sc_should_use_normal_mode once
Historically snap-confine used two modes of operation. Initially on core
systems snap-confine would unshare the mount namespace and preform some
alterations but would otherwise run with the contents of the initial
mount namesapce. On classic systems snap-confine would pivot into the
base snap (core) and perform some more alterations.

With the introduction of multiple boot bases, per-snap bases some of
that logic seems rusty. To make it less confusing what is going on
reduce the number of places that determine if legacy (sans-pivot) or
normal mode (with pivot) to exactly one.

There is more cleanup that can be done but it will require making the
invocation structure something we can pass around to more functions.
This will be attempted in subsequent patches.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>