cmd/snap-confine: move is_normal_mode to sc_invocation
This allows us to almost start passing sc_invocation instead of a
plethora of arguments to various functions inside ns-support and
mount-support modules.

This also makes it a tiny bit problematic because is_normal_mode is a
boolean that is only set inside the non-classic code path. This is
correct for now but should be carefully changed later.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>