interfaces/seccomp: snap-seccomp is at the same location as snapd
The snap-seccomp tool is expected to be found at the same location as snapd. Use
a sane default only when everything else fails.

This fixed seccomp backend initialization in a bootstrap environment, where
snapd/core snap can be mounted at some random location, but snapd & tools are
not available inside the rootfs yet.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>