cmd/snap-confine: drop unused dependency on libseccomp
Historically snap-confine would use libseccomp to compile and attach a
seccomp profile on startup of each new snap application or hook.

For a number or releases however this has not been the case. Instead the
seccomp profile is compiled with snap-seccomp (which does link to
libseccomp) and is stored on disk in /var/lib/snapd/seccomp/bpf.  Since
the change all that snap-confine has do to is to load the compiled
profile from disk and pass it to a system call for attachment. Therefore
the whole dependency as well as the associated autoconf/automake
boilerplate can be discarded.

With this change we do exactly what we did before but in there is one
less variant of snap-confine that may be built that would require
testing.

packaging/ubuntu: stop passing --enable-static-libseccomp

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>