cmd/snap-confine: move ubuntu-core fallback checks
Historically snap-confine would use the "ubuntu-core" base snap by
default (before base-snaps were officially recognized). During the
development of the ubuntu-core -> core migration feature in snapd,
snap-confine grew the ability to detect that new snap-confine (from
"new" snapd) is running on a system that has not yet completed the
migration process. When this happens snaps that were asked to run on top
of the "core" snap, which is not installed yet, would transparently
switch over to use "ubuntu-core" instead.

The associated logic was placed in mount-support.c, just as we were
preparing the bootstrap process of the new mount namespace.
Conveniently, a local variable was changed from "core" to "ubuntu-core"
and the associated spread test was passing.

Recently we wanted to add another such fallback, from "core16" to "core"
but due to much larger exposure (that fallback would be running on many
machines all the time) we put it under extra scrutiny and found an
inconsistency: the local modification was seen only by a part of
snap-confine. Some of the logic was making decisions based on the name
"core" while the rest would see "ubuntu-core". The similar problem would
be repeated for "core16".

With the introduction of sc_invocation we can now address this problem.
The invocation grew a helper function that can apply the core ->
ubuntu-core fallback globally. The fallback is applied correctly in the
initial mount namespace (that of the init process), before we take the
crucial decision whether normal (that is pivot) or legacy mode (without
pivot) should be used.

The fallback logic stays the same, except that now
sc_should_use_normal_mode() also handles the name "ubuntu-core" for
obvious reasons, so that the fallback logic stays the same.

With this change we can now introduce additional fallback logic safely.

One of the tests was updated to measure an updated error string.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>