cmd/snap-update-ns: implement TransitionalProfileUpdate.NeededChanges
This implements the lowest hanging fruit on the path to the refactoring.
NeededChanges is a member of the interface to allow conjuring desired
behavior of this complex part separately of the rest of the update logic.

It is also possible that nvidia transition will require a hand-crafted
implementation of NeededChanges when running in the system-wide, per-snap
mount namespace.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>