cmd/snap-update-ns: add testProfileUpdate
testProfileUpdate implements MountProfileUpdate and is useful for
testing as the default implementation of all the interface methods
does nothing but can be easily overridden on a per-test-case basis.

This mostly removes the need of global function replacement (mocking)
with the exception of some cases that don't yet expose a mount profile
update interface as argument.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>