cmd/snap-update-ns: use common load/save for SystemProfileUpdate
This removes first half of the duplicate logic, the other half being
the per-user mount profile logic, allowing for more unification
between the two algorithms (system and user).

We non also need to initialize private fields of SystemProfileUpdate
so there is a new function NewSystemProfileUpdate that allows this.

Some of the tests also return to a shape closer to the initial one,
since before the refactoring, due to the replacement of
testProfileUpdate with SystemProfileUpdate which implements the real
logic. This is better for a refactoring since tests visibly are nearly
the same as before.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>