cmd/snap-update-ns: move locking to the common layer
The locking was, so far, only performed for system-wide updates.
This property is retained since this is simply a refactoring at this
stage. Locking is now fully tested and documented.

The common layer is now aware if snap-update-ns is invoked from
snap-confine or from snapd. When invoked from snapd locking is handled
entirely by snap-update-ns. When invoked from snap-confine (during
the namespace bootstrap process) locking is verified. In both cases
the snap processes are frozen.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>