cmd/snap-update-ns: make apply{User,System}Fstab identical
This mechanical step is done to ensure that subsequent changes
did not break the system. Note that the per-user mount profile update
logic now performs locking (which is no-op), loads the current profile
(which is empty), synthesizes additional filesystem assumptions based
on that (empty) profile (so no new assumptions). Lastly the updated
current profile is saved (which is a no-op).

In short, this makes user and system profile application use identical
code without changing anything in practice.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>