cmd/snap-update-ns: add tests for executeMountProfileUpdate
The tests cover all the logic and branching avenues. The tests are
expanding upon the tests in main_test.go which date back from the first
implementation of the snap-update-ns program. The new tests use the more
sophisticated mocking logic to to concentrate each test on a specific
part of tested logic.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>