snapstate,state: add TaskSet.AddAllWithEdges() and use in doUpdate
The current code in doUpdate works on a bunch of TaskSets like
alias-updates, snap-updates and those are merged into a single
TaskSet in the end. With that the TaskSetEdge information gets
lost. This edge is import for the Remodel() code so we need
to preserve it.

This PR adds a new TaskSet.AddAllWithEdges() and uses it in
doUpdate() to keep the edges informations in the flat TaskSet.