snapstate: add new NoReRefresh flag and use in Remodel()
During a Remodel() we cannot assume the network is available
during the entire operation. When we generate a refresh change
in Remodel (e.g. because the kernel track changes) we are
currently generating a "rerefresh" task. This task will try
to reach the network and (worse) add a bunch of extra tasks
into the remodel change. This will break our carefully constructed
wait chain.

This PR adds a new NoReRefresh flag that will skip adding the
rerefresh task. It also adds a manager_test.go test that runs
a Remodel() with a kernel track change.