cmd/libsnap,osutil: fix parsing of mountinfo
The SUSE security team has performed an audit of snapd and found that
parsing mountinfo is vulnerable mishandling to certain crafted input.
Specifically anything that is regarded as space by sscanf for the
purpose of delimiting %s but is not regarded as escape-worthy space by
the kernel will confuse the parser.

The fix is applied to both the C and Go implementation of the parser.

The C patch changes two things in one go, because the switch from sscanf
made it easy. The code now handles kernel escape sequences for the four
characters that the kernel escapes (space, tab, newline and forward
slash). In addition the code no longer treats anything but space as
actual field delimiter.

This is now more in line with how the kernel behaves. There are
additional tests for some interesting cases.

Lastly, some size_t vs int fixes that were not reported by gcc before
were addressed. The code was also re-verified with valgrind.

The Go patch is smaller, we split on only space, not any white space
characters. The Go side always handled escape logic so no change is
needed there.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>
Related-To: https://bugzilla.suse.com/show_bug.cgi?id=1127368