cmd/snap-confine: umount scratch dir using UMOUNT_NOFOLLOW
During the start-up sequence, snap-confine performs some interesting
operations to prepare for pivot root. This involves setting up a root
filesystem in scratch directory in /tmp. After the pivot that directory
is unmounted (it was earlier mounted as an unbindable bind-mount).

All in all this is not a security risk of any kind but due diligence
similar to O_NOFOLLOW for openat(2) and similar.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>
Related-To: https://bugzilla.suse.com/show_bug.cgi?id=1127368