boot: add flag file "force-kernel-extraction"
Some kernels will need be extracted outside of the kernel squashfs
into /boot even when running on grub. One use-case for this is
fully encrypted systems. To support this feature this PR adds a flag
file "force-kernel-extraction" that will force the extraction of
the kernel and initramfs into /boot/$kernelname_$revno/.

The grub.cfg in the custom gadget needs to be adjusted for this
too.