devicestate: deal correctly with the "required" flag on Remodel
When doing a devicestate.Remodel() we need to ensure that all
newly required snaps are marked in the state as "Required: true"
and that we remove the "Required" flag from snaps that are no
longer required in the new model.

This PR does that and adds a spread test to ensure this works
correctly.