overlord/snapstate: reduce refresh delegate to callback
The soft/hard refresh check was implemented using a delegate that would
decide if an app or a hook can run while a refresh is in progress. The
answer for hooks was always "false" so we can reduce the complexity by
discarding the delegate and providing a single callback for apps.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>