cmd/snap-update-ns: remove MountProfileUpdate.NeededChanges
Just like in the patch before, the MountProfileUpdate exists for
decoupling the complexity behind each method from the algorithm that
operates on the interface implementations which in turn simplifies
testing.

The NeededChanges is a purely functional algorithm that doesn't need to
be usually mocked in tests. The case when it is mocked, in some of my
future work, is done only as a shortcut to avoid more preparation of the
load desired / load current dance. In case it is strictly needed for
clarity we can introduce MockNeededChanges

As such, remove the method from the interface and use the NeededPerform
function directly.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>