cmd/snap-update-ns: rename MountProfileUpdate{,Context}
The MountProfileUpdate interface was designed for decoupling an
algorithm from a set of complex functions it invokes, which aids in
testing. The recent trimming of that interface allows us to give it
a proper name. There are two implementations of the interface, the user
and system. By calling it a mount profile update *context* we can
treat it as a real interface, more than a testing helper.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>