cmd/snap-confine: prevent cwd restore permission bypass
It was brought to our attention that specially crafted current working
directory can be used to bypass permissions checks on a multi-user
system. The principal reason for that was that snap-confine, during the
course of execution, changes the current working directory, either
explicitly or implicitly (e.g. via setns) and then tries to restore the
original working directory while executing with root privileges.

The recently introduced central handling of umask is now extended to
handle current working directory. The code replaces individual
implementation in several other places in the codebase.

The new implementation is better than the original in two ways.  First
of all, it works based on a different principle which is discussed
below. Second of all, it runs late, after the original user and group
IDs have been restored (that is, it no longer runs as privileged user).

The new method has one of three possible outcomes:

1) The original working directory cannot be represented in the execution
environment. The process is moved to /var/lib/snapd/void

2) The original working directory can be represented in the execution
environment but it now points to a different inode. The new inode is
used and a filesystem permission check is performed by the kernel.

3) The original working directory can be represented in the execution
environment and points to the same inode. The same inode is used.

Two tests were affected by this change as they were running commands
inside a LXD container as the root user, with the initial working
directory set to /root, then using su to switch to the ubuntu user.
Those commands were changed to go to /home/ubuntu instead.

To show this all in operation I've added a new test that verifies each
of the variant cases.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>