tests/lib/pkgdb: allow downgrade when installing packages in openSUSE
Packages may be downgraded in the main repositories, which would be picked up
next time we ran `zypper dup` and applied locally. However we only update the
images periodically, in the meantime, when downgrades affect packages we need or
have installed, `zypper in` may stop with the prompt asking the user about
either breaking the installed packages or allowing downgrades. Passing
--allow-downgrade explicitly will make the installation proceed.

Hopefully the spread jobs on openSUSE would fail less often too.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>