cmd/snap-confine: move to void on permission errors
When the invoking user has no permission to stay in a directory (e.g. a
symlink attack in /tmp or a su invocation from /root) then instead of
failing the whole command chain, move the invoked process to the void
directory.

The void directory is empty and has no permissions to do anything
inside. The set of affected tests were changed not to require a
directory change anymore. In addition the pwd test was expanded to cover
this specific situation.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>