packaging,tests: ensure that /var/lib/snapd/void is mode 111
The void directory cannot be 000 anymore. When snap-confine chooses to
restore the orignal working directory inside the mount namespace it must
do so without abusing the root power it wields. As such it drops extra
permissions and moves to the desired directory as the invoking user.

This can fail for lack of permissions to access said directory. When
this happens, instead of aborting execution, snap-confine moves to the
special fallback directory, the void.

This raises the problem that it cannot access the void directory as
originally designed (with mode 000) so something must yield. We chose
to yield the permissions by changing them to 111, that is, anyone can
browse but nothing else.

The change is done across many layers of packaging and test preparation
but even that is insufficient. What we noticed is that the originally
desired permissions were not really applied.

I will separately address this in core and core18 so that the directory
exists and with correct permissions but for now this should unblock the
effort.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>